requirements.txt
Flask==3.0.0
werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
